got rid off listOfStates entirely; way faster!
it turns out that there's no need to keep a list of all possible states around, since the numberOfStates already encodes that information by being the number that it is. I had actually unwittingly removed most reference to listOfStates in an earlier edit, but today I realized that nothing was calling it, and all it was doing was sitting bloating and slowing the whole program.

now the program is drastically faster. in particular, the initial load time has vanished for reasonably low input parameters. the program also uses something like 90% less memory. further, it's now able to handle much greater input parameters; in particular, it can now handle colors > 9.

TODO:
* documentation!!!
* there still seems to be a bug where the pattern is incredibly boring is the number is even and > 2. I have no clue what is causing that.